Move AvatarData to components/Avatar
