+++ site/src/modules/builds/BuildAvatar/BuildAvatar.tsx
@@ -2,8 +2,8 @@ import { css, cx } from "@emotion/css";
 import { useTheme } from "@emotion/react";
 import Badge from "@mui/material/Badge";
 import type { WorkspaceBuild } from "api/typesGenerated";
-import { BuildIcon } from "components/BuildIcon/BuildIcon";
 import { Avatar, type AvatarProps } from "components/Avatar/Avatar";
+import { BuildIcon } from "components/BuildIcon/BuildIcon";
 import { useClassName } from "hooks/useClassName";
 import type { FC } from "react";
 import { getDisplayWorkspaceBuildStatus } from "utils/workspace";
